minor fix related to initSpectrumComponents
