+++ ecc/blocks/form-handler/form-handler.js
@@ -3,20 +3,8 @@ import {
   signIn,
 } from '../../scripts/utils.js';
 
-import { LIBS } from '../../scripts/scripts.js';
 import { initProfileLogicTree } from '../../scripts/profile.js';
-import { buildLoadingScreen, buildECCForm, SPECTRUM_COMPONENTS } from './form-handler-helper.js';
-
-async function initSpectrumComponents() {
-  const miloLibs = LIBS;
-  const promises = Array.from(SPECTRUM_COMPONENTS).map(async (component) => {
-    await import(`${miloLibs}/features/spectrum-web-components/dist/${component}.js`);
-  });
-  await Promise.all([
-    import(`${miloLibs}/deps/lit-all.min.js`),
-    ...promises,
-  ]);
-}
+import { buildLoadingScreen, buildECCForm, initSpectrumComponents } from './form-handler-helper.js';
 
 export default async function init(el) {
   buildLoadingScreen(el);
